renamed all source from .js to .ts
