+++ src/classes/patterns.ts
@@ -1,4 +1,4 @@
-import { DataBlock } from "./datablock.mjs";
+import { DataBlock } from "./datablock.js";
 
 export class Pattern extends DataBlock {
   constructor(data) {
