+++ src/ui/slider.ts
